Adjusting the example in specification to be less confusing.
The original example included a comment after the “uncommented” `…` in order to explain that it *did* reference real-but-unmentioned files.

Given that we’re discussing *uncommented* behavior, this seems well-intentioned but unnecessarily confusing to a casual reader…so I removed the comment from the example, and added an explanatory `Note:` communicating the same assumptions.